flask_app/requirements.txt
flask>=3.1.0
flask-cors>=5.0.0
flask-caching>=2.3.0
PyJWT>=2.10.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
authlib>=1.3.0
requests>=2.31.0
pandas
openpyxl
scipy
numpy
